Refactor to use Bson serialization for everything (dependancy: uk.dioxic.kotlinx.serialization.bson)
