+++ cli/src/main/kotlin/uk/dioxic/helios/cli/commands/Benchmark.kt
@@ -29,10 +29,10 @@ import uk.dioxic.helios.execute.format.format
 import uk.dioxic.helios.execute.format.toFormatString
 import uk.dioxic.helios.execute.resources.ResourceRegistry
 import uk.dioxic.helios.execute.resources.mongoClient
-import uk.dioxic.helios.generate.codecs.TemplateCodec
 import kotlin.time.ExperimentalTime
 import kotlin.time.measureTime
 import uk.dioxic.helios.execute.model.Benchmark as ExBenchmark
+import uk.dioxic.helios.generate.codecs.DocumentCodec as HeliosDocumentCodec
 
 class Benchmark : CliktCommand(help = "Execute Benchmark") {
     init {
@@ -56,7 +56,7 @@ class Benchmark : CliktCommand(help = "Execute Benchmark") {
         val mcs = MongoClientSettings.builder()
             .applyAuthOptions(authOptions)
             .applyConnectionOptions(connOptions)
-            .codecRegistry(TemplateCodec.defaultRegistry)
+            .codecRegistry(HeliosDocumentCodec.defaultRegistry)
             .build()
