+++ cli/src/main/kotlin/uk/dioxic/helios/cli/commands/Load.kt
@@ -36,12 +36,12 @@ import uk.dioxic.helios.execute.resources.ResourceRegistry
 import uk.dioxic.helios.execute.resources.mongoClient
 import uk.dioxic.helios.generate.Template
 import uk.dioxic.helios.generate.buildTemplate
-import uk.dioxic.helios.generate.codecs.TemplateCodec
 import kotlin.math.min
 import kotlin.math.roundToInt
 import kotlin.time.DurationUnit
 import kotlin.time.ExperimentalTime
 import kotlin.time.measureTime
+import uk.dioxic.helios.generate.codecs.DocumentCodec as HeliosDocumentCodec
 
 class Load : CliktCommand(help = "Load data directly into MongoDB") {
     init {
@@ -74,7 +74,7 @@ class Load : CliktCommand(help = "Load data directly into MongoDB") {
         val mcs = MongoClientSettings.builder()
             .applyAuthOptions(authOptions)
             .applyConnectionOptions(connOptions)
-            .codecRegistry(TemplateCodec.defaultRegistry)
+            .codecRegistry(HeliosDocumentCodec.defaultRegistry)
             .build()
 
         val amendedBatchSize = min(batchSize, number.toInt())
