+++ generate/src/main/kotlin/uk/dioxic/helios/generate/Hydration.kt
@@ -15,7 +15,7 @@ fun Map<String, *>.hydrate(): Map<String, *> =
 
 context (OperatorContext)
 fun Template.hydrate(): Map<String, *> =
-    hydrate(this) as Map<String, *>
+    hydrate(execution) as Map<String, *>
 
 private val rootOperatorKey = getOperatorKey<RootOperator>()
