+++ generate/src/main/kotlin/uk/dioxic/helios/generate/Maps.kt
@@ -2,6 +2,7 @@ package uk.dioxic.helios.generate
 
 import org.bson.BsonDocument
 import org.bson.BsonValue
+import org.bson.Document
 
 fun Map<String, Any?>.flatten(separator: Char = '.', leafOnly: Boolean = false) =
     mutableMapOf<String, Any?>().also {
@@ -13,10 +14,8 @@ fun BsonDocument.flatten(separator: Char = '.', leafOnly: Boolean = false) =
         flatten(it, this, separator, leafOnly)
     }.toMap()
 
-fun <V> templateOf(pair: Pair<String, V>): Template = Template(mapOf(pair))
-
-fun <V> templateOf(vararg pairs: Pair<String, V>): Template =
-    Template(mapOf(*pairs))
+fun Map<String, Any?>.toDocument() =
+    Document(this)
 
 @Suppress("UNCHECKED_CAST")
 private fun <T> flatten(
@@ -57,3 +56,7 @@ private fun getKey(prefix: String, separator: Char, suffix: String) =
     } else {
         "$prefix$separator$suffix"
     }
+
+inline fun <reified O: Operator<*>> MutableMap<String, Any>.putOperator(key: String) {
+    put(key, getOperatorKey<O>())
+}
