+++ generate/src/main/kotlin/uk/dioxic/helios/generate/codecs/WrappedCodecProvider.kt
@@ -3,14 +3,14 @@ package uk.dioxic.helios.generate.codecs
 import org.bson.codecs.Codec
 import org.bson.codecs.configuration.CodecProvider
 import org.bson.codecs.configuration.CodecRegistry
-import uk.dioxic.helios.generate.Operator
+import uk.dioxic.helios.generate.Wrapped
 
-class OperatorExecutionCodecProvider: CodecProvider {
+class WrappedCodecProvider: CodecProvider {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : Any?> get(clazz: Class<T>, registry: CodecRegistry): Codec<T>? {
-        if (Operator::class.java.isAssignableFrom(clazz)) {
-            return OperatorExecutionCodec(registry) as Codec<T>
+        if (Wrapped::class.java.isAssignableFrom(clazz)) {
+            return WrappedCodec(registry) as Codec<T>
         }
         return null
     }
