+++ generate/src/test/kotlin/uk/dioxic/helios/generate/TemplateSerializationTests.kt
@@ -4,10 +4,11 @@ import io.kotest.assertions.throwables.shouldNotThrowAny
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.should
+import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.serialization.bson.Bson
 import kotlinx.serialization.bson.buildBsonDocument
-import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.bson.decodeFromBsonDocument
 import kotlinx.serialization.encodeToString
 import org.bson.Document
 import uk.dioxic.helios.generate.operators.ChooseOperator
@@ -17,7 +18,7 @@ class TemplateSerializationTests : FunSpec({
 
     val bson = Bson { prettyPrint = true }
 
-    val jsonObject = buildBsonDocument {
+    val definition = buildBsonDocument {
         putOperatorObject<ChooseOperator>("color") {
             putBsonArray("from") {
                 addAll(listOf("blue", "red", "green"))
@@ -28,31 +29,29 @@ class TemplateSerializationTests : FunSpec({
             putOperator<ChooseOperator>("city", listOf("london", "new york", "dulwich"))
         }
     }
-    val jsonString = bson.encodeToString(jsonObject)
-    println(jsonString)
 
     test("decode") {
-        bson.decodeFromString<Template>(jsonString).should {
-            it.shouldBeInstanceOf<Document>()
-            it["color"].shouldBeInstanceOf<ChooseOperator>()
-            it["height"].shouldBeInstanceOf<IntOperator>()
-            it["address"].should { addr ->
-                addr.shouldBeInstanceOf<Document>()
-                addr["city"].shouldBeInstanceOf<ChooseOperator>()
+        bson.decodeFromBsonDocument<Template>(definition).should {template ->
+            template.execution.should {
+                it["color"].shouldBeInstanceOf<ChooseOperator>()
+                it["height"].shouldBeInstanceOf<IntOperator>()
+                it["address"].should { addr ->
+                    addr.shouldBeInstanceOf<Document>()
+                    addr["city"].shouldBeInstanceOf<ChooseOperator>()
+                }
             }
+            template.definition shouldBe definition
         }
     }
 
     test("encode without definition should fail") {
-        val template = Template(emptyMap<String, String>())
-
         shouldThrow<IllegalArgumentException> {
-            bson.encodeToString(template)
+            println(bson.encodeToString(Template(Document())))
         }
     }
 
     test("encode with definition should succeed") {
-        val template = Template(emptyMap<String, String>(), jsonObject)
+        val template = Template(Document(), definition)
 
         shouldNotThrowAny {
             bson.encodeToString(template)
